requirements.txt
Flask-Cors
Flask
requests
azure-ai-formrecognizer
azure-core
PyPDF2